:truck: chore(api) move `PrescripberRoleReporitory` to `src/`
