+++ api/src/shared/application/usecases/checkAuthorizationToAccessCampaign.js
@@ -1,5 +1,5 @@
-import * as prescriberRoleRepository from '../../../../lib/infrastructure/repositories/prescriber-role-repository.js';
-import { CampaignAuthorization } from '../../../../src/shared/application/pre-handlers/CampaignAuthorization.js';
+import * as prescriberRoleRepository from '../../infrastructure/repositories/prescriber-role-repository.js';
+import { CampaignAuthorization } from '../pre-handlers/CampaignAuthorization.js';
 
 const execute = async function ({ userId, campaignId }) {
   const prescriberRole = await prescriberRoleRepository.getForCampaign({ userId, campaignId });
