+++ api/src/shared/application/usecases/checkAuthorizationToManageCampaign.js
@@ -1,4 +1,4 @@
-import * as prescriberRoleRepository from '../../../../lib/infrastructure/repositories/prescriber-role-repository.js';
+import * as prescriberRoleRepository from '../../infrastructure/repositories/prescriber-role-repository.js';
 import { CampaignAuthorization } from '../pre-handlers/CampaignAuthorization.js';
 
 const execute = async function ({ userId, campaignId }) {
