+++ api/tests/shared/integration/infrastructure/repositories/prescriber-role-repository_test.js
@@ -1,5 +1,5 @@
-import * as prescriberRoleRepository from '../../../../lib/infrastructure/repositories/prescriber-role-repository.js';
-import { databaseBuilder, expect } from '../../../test-helper.js';
+import * as prescriberRoleRepository from '../../../../../src/shared/infrastructure/repositories/prescriber-role-repository.js';
+import { databaseBuilder, expect } from '../../../../test-helper.js';
 
 describe('Integration | Repository | prescriber-role-repository', function () {
   describe('#getForCampaign', function () {
